chore: remove all shared plots
