--- src/psycop/common/model_evaluation/binary/time/test_absolute_plots.py
@@ -1,23 +1,0 @@
-from typing import Literal
-
-import pytest
-from psycop.common.model_evaluation.binary.time.absolute_plots import (
-    plot_metric_by_absolute_time,
-)
-from psycop.common.model_evaluation.utils import TEST_PLOT_PATH
-from psycop.common.model_training.training_output.dataclasses import EvalDataset
-
-
-@pytest.mark.parametrize(
-    "bin_period",
-    ["Q", "Y"],
-)
-def test_plot_performance_by_calendar_time(
-    subsampled_eval_dataset: EvalDataset,
-    bin_period: Literal["M", "Q", "Y"],
-):
-    plot_metric_by_absolute_time(
-        eval_dataset=subsampled_eval_dataset,
-        bin_period=bin_period,
-        save_path=TEST_PLOT_PATH / f"test_{bin_period}.png",  # type: ignore
-    )
