+++ src/psycop/common/model_evaluation/binary/time/test_timedelta_data.py
@@ -1,7 +1,9 @@
 from psycop.common.model_evaluation.binary.time.timedelta_data import (
+    create_sensitivity_by_time_to_outcome_df,
     get_sensitivity_by_timedelta_df,
     get_timedelta_series,
 )
+from psycop.common.model_training.training_output.dataclasses import EvalDataset
 from psycop.common.test_utils.str_to_df import str_to_df
 
 
@@ -55,3 +57,12 @@ def test_sensitivity_by_timedelta():
 
     ci_lower = sensitivity_df["ci"][0][0][0]
     assert ci_lower == 1.0
+
+
+def test_get_sens_by_time_to_outcome_df(synth_eval_dataset: EvalDataset):
+    create_sensitivity_by_time_to_outcome_df(
+        eval_dataset=synth_eval_dataset,
+        outcome_timestamps=synth_eval_dataset.outcome_timestamps,
+        prediction_timestamps=synth_eval_dataset.pred_timestamps,
+        desired_positive_rate=0.5,
+    )
