+++ src/psycop/projects/t2d/paper_outputs/model_description/performance/sensitivity_by_time_to_event_pipeline.py
@@ -1,21 +1,12 @@
-from psycop.common.model_evaluation.binary.time.timedelta_plots import (
-    plot_sensitivity_by_time_to_event,
-)
 from psycop.projects.t2d.paper_outputs.config import EVAL_RUN, FIGURES_PATH
 from psycop.projects.t2d.utils.best_runs import Run
 
 
 def sensitivity_by_time_to_event(run: Run):
-    eval_ds = run.get_eval_dataset()
+    run.get_eval_dataset()
 
     plot_path = FIGURES_PATH / "sensitivity_by_time_to_event.png"
-    plot_sensitivity_by_time_to_event(
-        eval_dataset=eval_ds,
-        bins=range(0, 37, 6),
-        positive_rates=[0.05, 0.03, 0.01],
-        bin_unit="M",
-        save_path=plot_path,
-    )
+    # TODO: Create plot
     print(f"Saving sensitivity by time to event plot to {plot_path}")
