+++ src/psycop/projects/t2d/paper_outputs/model_description/robustness/robustness_by_age.py
@@ -1,7 +1,5 @@
-from psycop.common.model_evaluation.binary.subgroups.age import (
-    plot_roc_auc_by_age,
-)
-from psycop.projects.t2d.paper_outputs.config import EVAL_RUN, ROBUSTNESS_PATH
+from psycop.common.model_evaluation.binary.subgroup_data import get_auroc_by_input_df
+from psycop.projects.t2d.paper_outputs.config import EVAL_RUN
 from psycop.projects.t2d.utils.best_runs import Run
 
 
@@ -9,11 +7,16 @@ def roc_auc_by_age(run: Run):
     print("Plotting AUC by age")
     eval_ds = run.get_eval_dataset()
 
-    plot_roc_auc_by_age(
+    get_auroc_by_input_df(
         eval_dataset=eval_ds,
+        input_values=eval_ds.age,  # type: ignore
+        input_name="age",
         bins=[18, *range(20, 80, 10)],
-        save_path=ROBUSTNESS_PATH / "auc_by_age.png",
+        bin_continuous_input=True,
+        confidence_interval=True,
     )
+
+    # TODO: Create plot function
 
 
 if __name__ == "__main__":
