+++ src/psycop/projects/t2d/paper_outputs/model_description/robustness/robustness_by_n_hba1c.py
@@ -2,10 +2,7 @@ from collections.abc import Sequence
 from pathlib import Path
 from typing import Optional, Union
 
-from psycop.common.model_evaluation.base_charts import (
-    plot_basic_chart,
-)
-from psycop.common.model_evaluation.binary.subgroups.base import create_roc_auc_by_input
+from psycop.common.model_evaluation.binary.subgroup_data import get_auroc_by_input_df
 from psycop.common.model_training.training_output.model_evaluator import EvalDataset
 from psycop.projects.t2d.paper_outputs.config import EVAL_RUN, ROBUSTNESS_PATH
 from psycop.projects.t2d.utils.best_runs import Run
@@ -13,10 +10,10 @@ from psycop.projects.t2d.utils.best_runs import Run
 
 def plot_performance_by_n_hba1c(
     eval_dataset: EvalDataset,
-    save_path: Optional[Path] = None,
+    save_path: Optional[Path] = None,  # noqa: ARG001
     bins: Sequence[float] = (18, 25, 35, 50, 70),
     bin_continuous_input: Optional[bool] = True,
-    y_limits: Optional[tuple[float, float]] = (0.5, 1.0),
+    y_limits: Optional[tuple[float, float]] = (0.5, 1.0),  # noqa: ARG001
 ) -> Union[None, Path]:
     """Plot bar plot of performance (default AUC) by age at time of prediction.
 
@@ -32,7 +29,7 @@ def plot_performance_by_n_hba1c(
     """
     col_name = "eval_hba1c_within_9999_days_count_fallback_nan"
 
-    df = create_roc_auc_by_input(
+    df = get_auroc_by_input_df(
         eval_dataset=eval_dataset,
         input_values=eval_dataset.custom_columns[col_name],  # type: ignore
         input_name=col_name,
@@ -41,23 +38,11 @@ def plot_performance_by_n_hba1c(
         confidence_interval=True,
     )
 
-    ci = df["ci"].tolist()
+    df["ci"].tolist()
 
-    sort_order = sorted(df[f"{col_name}_binned"].unique())
+    sorted(df[f"{col_name}_binned"].unique())
 
-    return plot_basic_chart(
-        x_values=df[f"{col_name}_binned"],  # type: ignore
-        y_values=df["metric"],  # type: ignore
-        x_title="HbA1c measurements before visit",
-        y_title="AUC",
-        sort_x=sort_order,
-        y_limits=y_limits,
-        plot_type=["scatter", "line"],
-        bar_count_values=df["n_in_bin"],
-        bar_count_y_axis_title="Number of visits",
-        confidence_interval=ci,
-        save_path=save_path,
-    )
+    # TODO: Implement plot function
 
 
 def plot_auroc_by_n_hba1c(run: Run):
