+++ src/psycop/projects/t2d/paper_outputs/model_description/robustness/robustness_by_sex.py
@@ -1,16 +1,12 @@
-from psycop.common.model_evaluation.binary.subgroups.sex import plot_roc_auc_by_sex
-from psycop.projects.t2d.paper_outputs.config import EVAL_RUN, ROBUSTNESS_PATH
+from psycop.projects.t2d.paper_outputs.config import EVAL_RUN
 from psycop.projects.t2d.utils.best_runs import Run
 
 
 def roc_auc_by_sex(run: Run):
     print("Plotting AUC by sex")
-    eval_ds = run.get_eval_dataset(custom_columns=["is_female"])
+    run.get_eval_dataset(custom_columns=["is_female"])
 
-    plot_roc_auc_by_sex(
-        eval_dataset=eval_ds,
-        save_path=ROBUSTNESS_PATH / "auc_by_sex.png",
-    )
+    # TODO: Create plot
 
 
 # %%
